requirements.txt
Django==5.2.9
google-generativeai==0.8.2
python-dotenv==1.0.0
whitenoise==6.6.0
gunicorn==21.2.0